+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Feature] Create generic `OverlayWrapper` component that intelligently positions an overlay on a page relative to it's children. (#572)
+- [Feature] Create simple `Popover` component. (#462)
+- [Release] Import a new SCSS file for Tether, a third party library used to position popovers.
 
 ## [17.5.1][17.5.1] - 2016-09-28
 ### Fixed